+++ packages/deepsea-tools/src/utils/readExcel.ts
@@ -1,59 +1,11 @@
-import { ParsingOptions, Sheet2JSONOpts, read, utils } from "xlsx"
-
-export type { ParsingOptions, Sheet2JSONOpts } from "xlsx"
+import { ReadSheetParams, readSheet } from "./readSheet"
 
 /**
- * 工作簿的数据
+ * @deprecated 请使用 ReadSheetParams 代替
  */
-export type WorkBookData<Sheet = WorkSheetData> = {
-    [sheetName: string]: Sheet
-}
+export type ReadExcelParams = ReadSheetParams
 
 /**
- * 工作表的数据
+ * @deprecated 请使用 readSheet 代替
  */
-export type WorkSheetData = WorkSheetRowData[]
-
-/**
- * 单元格的值
- */
-export type WorkSheetCellValue = Date | string | number | boolean | undefined
-
-/**
- * 工作表的数据
- */
-export type WorkSheetRowData = {
-    /**
-     * 第几行
-     */
-    __rowNum__: number
-    /**
-     * key 是表头，value 是单元格的值
-     */
-    [columnName: string]: WorkSheetCellValue
-}
-
-export interface ReadExcelParams {
-    buffer: ArrayBuffer
-    /**
-     * cellDates 将会默认开启
-     */
-    parsingOptions?: ParsingOptions
-    sheetToJsonOptions?: Sheet2JSONOpts
-}
-
-/**
- * 读取工作簿
- * @param buffer 工作簿的二进制数据
- * @returns 工作簿的数据
- */
-export function readExcel<Sheet = WorkSheetData>(bufferOrParams: ArrayBuffer | ReadExcelParams): WorkBookData<Sheet> {
-    const { buffer, parsingOptions, sheetToJsonOptions } =
-        bufferOrParams instanceof ArrayBuffer ? ({ buffer: bufferOrParams } as ReadExcelParams) : bufferOrParams
-    const wb = read(buffer, { cellDates: true, ...parsingOptions })
-    const result = wb.SheetNames.reduce((acc, item) => {
-        acc[item] = utils.sheet_to_json(wb.Sheets[item], sheetToJsonOptions) as any
-        return acc
-    }, {} as WorkBookData<Sheet>)
-    return result
-}
+export const readExcel = readSheet